ManhwaOCR/requirements.txt
flask
easyocr
Pillow
torch
transformers
